section 10 finished
