+++ recipes/tests/test_recipe_home_view.py
@@ -1,0 +1,46 @@
+from django.urls import reverse, resolve
+from recipes import views
+from .test_recipe_base import RecipeTestBase
+
+class RecipeHomeViewTest(RecipeTestBase):
+    def test_recipes_home_views_function_is_correct(self):
+        view = resolve(reverse('recipes:home'))
+        self.assertIs(view.func, views.home)
+
+    def test_recipe_home_view_returns_status_code_200_ok(self):
+        response = self.client.get(reverse('recipes:home'))
+        self.assertEqual(response.status_code, 200)
+
+    def test_recipe_home_view_loads_correct_template(self):
+        response = self.client.get(reverse('recipes:home'))
+        self.assertTemplateUsed(response, 'recipes/pages/home.html')
+
+    def test_recipe_home_template_shows_no_recipes_found_if_no_recipes(self):
+        response = self.client.get(reverse('recipes:home'))
+        self.assertIn(
+            'No recipes found',
+            response.content.decode('utf-8')
+            )
+    
+    def test_recipe_home_template_loads_recipes(self):
+        self.make_recipe()
+        response = self.client.get(reverse('recipes:home'))
+        content = response.content.decode('utf-8')
+        response_context_recipes = response.context['recipes']
+
+        # check if one recipe exists
+        self.assertIn('Recipe Title', content)
+        self.assertEqual(len(response_context_recipes), 1)
+
+    def test_recipe_home_template_dont_load_recipes_not_publisehd(self):
+        """Test recipes is publisehd False don't show"""
+        self.make_recipe(is_published=False)
+
+        response = self.client.get(reverse('recipes:home'))
+        content = response.content.decode('utf-8')
+
+        # check if one recipe exists
+        self.assertIn(
+            'No recipes found',
+            response.content.decode('utf-8')
+            )
