+++ recipes/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_list_or_404, get_object_or_404
 from django.http import Http404
+from django.db.models import Q
 from utils.recipes.factory import make_recipe
 from recipes.models import Recipe
 
@@ -42,8 +43,18 @@ def search(request):
 
     if not search_term:
         raise Http404()
+    
+    recipes = Recipe.objects.filter(
+        Q(
+        Q(title__icontains=search_term) |
+        Q(description__icontains=search_term),
+        is_published=True
+    )).order_by('-id')
+    #recipes = recipes.order_by('-id')
+    #recipes = recipes.filter(is_published=True)
 
     return render(request, 'recipes/pages/search.html', context={
         'page_title': f"Search for '{search_term}' | ",
-        'search_term': search_term
+        'search_term': search_term,
+        'recipes': recipes 
     })
